requirements.txt
numpy>=1.19.0
PyYAML>=6.0
opencv-python>=4.5.0
face-recognition>=1.3.0
Flask>=2.0.0
Flask-HTTPAuth>=4.5.0
picamera2>=0.3.9
libcamera>=0.0.5
python-dotenv>=0.19.0 